\name{FMrate}
\docType{data}
\alias{FMrate}
\title{Same-different judgments on frequency modulated tones}
\description{
  This data set consists of same-different judgments of three subjects on 9 frquency
  modulated tones.
}
\usage{data(FMrate)}
\format{A data frame of 7 variables and 26,820 observations.}
\keyword{datasets}

