\name{lightness}
\docType{data}
\alias{lightness}
\title{Same-different judgments on simple gray patches}
\description{
  This data set consists of same-different judgments of 9 simple gray
  patches (of luminance XXX) presented under constant illumination (XXX).
}
\usage{lightness}
\format{A data frame of 8 variables and 35,760 observations.}
\references{
  Umbach, N (2013). \emph{Dimensionality of the Perceptual Space of
  Achromatic Surface Colors.} Dissertation, University of Tuebingen.
}
\keyword{datasets}

