\name{FMrate}
\docType{data}
\alias{FMrate}
\title{Same-different judgments on frequency modulated tones}
\description{
  This data set consists of same-different judgments of 9 frquency
  modulated tones (XXX).
}
\usage{FMrate}
\format{A data frame of 7 variables and 26,820 observations.}
\references{
  Umbach, N (2013). \emph{Dimensionality of the Perceptual Space of
  Achromatic Surface Colors.} Dissertation, University of Tuebingen.
}
\keyword{datasets}

