\name{FMrate}
\docType{data}
\alias{FMrate}
\title{Same-different judgments on frequency modulated tones}
\description{
  This data set consists of same-different judgments of three subjects on 9
  frequency modulated tones.
}
\usage{data(FMrate)}
\format{data frame of 7 variables and 26,820 observations.}
\details{The 7 variables are named 'resp', 'RT', 'id', 's1', 's2', 'phi1',
and 'phi2'.
\preformatted{resp}{gives the response of a subject in each trial. Can
either be 's' for same or 'd' for different.}
\preformatted{RT}{reaction time from stimulus onset until reaction of
subject.}
\preformatted{id}{subject id.}
\preformatted{s1}{stimulus which was presented in the first observation area; in
this case: first.}
\preformatted{s2}{stimulus which was presented in the second observation
area.}
\preformatted{phi1}{phase of modulated tones was randomly drawn. Gives
phase shift for stimulus presented in the first observation area.}
\preformatted{phi2}{see above. Phase shift for stimulus presented in the
second observation area.}
}
\keyword{datasets}

