\name{lightness}
\docType{data}
\alias{lightness}
\title{Same-different judgments on gray patches}
\description{
  This data set consists of same-different judgments of four subjects on 9 gray
  patches (of luminance XXX) presented under constant illumination (XXX).
}
\usage{data(lightness)}
\format{A data frame of 8 variables and 35,760 observations.}
\source{
  Umbach, N (2013). \emph{Dimensionality of the Perceptual Space of
  Achromatic Surface Colors.} Dissertation, University of Tuebingen.
}
\keyword{datasets}

